SSE-76: Change id to int8 in db

+++ src/main/resources/config/liquibase/changelog/20221220100000_create_discount.sql
@@ -1,5 +1,5 @@
 create table discount (
-    id int NOT NULL,
+    id int8 NOT NULL,
     discount_code varchar(50) NOT NULL UNIQUE,
     discount_percentage float NOT NULL,
     PRIMARY KEY (id)
